Re-organize tests so that there is 1 test file per module
