+++ tests/test_api_utils_response.py
@@ -11,14 +11,10 @@
 #  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  * See the License for the specific language governing permissions and
 #  * limitations under the License.
-import os
 
 import pint
-import pytest
 
-from pims.api.utils.parameter import filepath2path, path2filepath
 from pims.api.utils.response import convert_quantity, response_list
-from pims.files.file import Path
 
 
 def test_response_list():
@@ -37,16 +33,3 @@ def test_convert_quantity():
 
     ureg = pint.UnitRegistry()
     assert convert_quantity(3 * ureg('cm'), 'meters') == 0.03
-
-
-@pytest.mark.parametrize("filepath", ("/abc", "abc", "abc/foo"))
-def test_filepath2path(app, settings, filepath):
-    assert str(filepath2path(filepath, settings)) == os.path.join(settings.root, filepath)
-
-
-@pytest.mark.parametrize("rootpath", ("/abc", "abc", "abc/foo/"))
-def test_path2filepath(app, settings, rootpath):
-    fake_settings = settings.copy()
-    fake_settings.root = rootpath
-    path = Path(rootpath) / "dir/file"
-    assert path2filepath(path, fake_settings) == "dir/file"
