+++ tests/test_processing_annotations.py
@@ -14,7 +14,8 @@
 
 from shapely.geometry import box
 
-from pims.processing.annotations import ParsedAnnotation, ParsedAnnotations
+from pims.processing.annotations import ParsedAnnotation, ParsedAnnotations, get_annotation_region
+from pims.processing.region import Region
 from pims.utils.color import Color
 
 
@@ -54,3 +55,31 @@ def test_annotation_list():
     assert al.is_stroke_grayscale is False
     assert al.is_grayscale is False
     assert al.bounds == (5, 20, 30, 200)
+
+
+def test_annotation_region():
+    class FakeImage:
+        def __init__(self, w, h):
+            self.width = w
+            self.height = h
+
+    al = ParsedAnnotations()
+    al.append(ParsedAnnotation(box(10, 20, 30, 40)))
+    assert get_annotation_region(
+        FakeImage(100, 100), al
+    ) == Region(20, 10, 20, 20)
+    assert get_annotation_region(
+        FakeImage(100, 100), al, context_factor=1.5
+    ) == Region(15, 5, 30, 30)
+
+    al = ParsedAnnotations()
+    al.append(ParsedAnnotation(box(10, 20, 30, 30)))
+    assert get_annotation_region(
+        FakeImage(100, 100), al, try_square=True
+    ) == Region(15, 10, 20, 20)
+
+    al = ParsedAnnotations()
+    al.append(ParsedAnnotation(box(20, 10, 30, 30)))
+    assert get_annotation_region(
+        FakeImage(100, 100), al, try_square=True
+    ) == Region(10, 15, 20, 20)
