+++ tests/test_utils_color.py
