+++ tests/test_utils_math.py
@@ -1,0 +1,20 @@
+#  * Copyright (c) 2020-2021. Authors: see NOTICE file.
+#  *
+#  * Licensed under the Apache License, Version 2.0 (the "License");
+#  * you may not use this file except in compliance with the License.
+#  * You may obtain a copy of the License at
+#  *
+#  *      http://www.apache.org/licenses/LICENSE-2.0
+#  *
+#  * Unless required by applicable law or agreed to in writing, software
+#  * distributed under the License is distributed on an "AS IS" BASIS,
+#  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  * See the License for the specific language governing permissions and
+#  * limitations under the License.
+
+from pims.utils.math import get_rationed_resizing
+
+
+def test_get_rationed_resizing():
+    assert get_rationed_resizing(50, 100, 200) == (50, 100)
+    assert get_rationed_resizing(0.5, 100, 200) == (50, 100)
